+++ src/AuraRouter.php
@@ -1,17 +1,25 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-aurarouter for the canonical source repository
- * @copyright Copyright (c) 2015-2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2015-2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-aurarouter/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\Expressive\Router;
 
 use Aura\Router\Route as AuraRoute;
 use Aura\Router\RouterContainer as Router;
 use Aura\Router\Rule\Path as PathRule;
-use Fig\Http\Message\RequestMethodInterface as RequestMethod;
 use Psr\Http\Message\ServerRequestInterface as Request;
+
+use function array_key_exists;
+use function array_merge;
+use function array_reduce;
+use function array_unique;
+use function explode;
+use function in_array;
 
 /**
  * Router implementation bridging the Aura.Router.
@@ -27,14 +35,6 @@ use Psr\Http\Message\ServerRequestInterface as Request;
  */
 class AuraRouter implements RouterInterface
 {
-    /**
-     * Implicit HTTP methods (should work for any route)
-     */
-    const HTTP_METHODS_IMPLICIT = [
-        RequestMethod::METHOD_HEAD,
-        RequestMethod::METHOD_OPTIONS,
-    ];
-
     /**
      * Map paths to allowed HTTP methods.
      *
@@ -63,8 +63,6 @@ class AuraRouter implements RouterInterface
      * If no Aura.Router instance is provided, the constructor will lazy-load
      * an instance. If you need to customize the Aura.Router instance in any
      * way, you MUST inject it yourself.
-     *
-     * @param null|Router $router
      */
     public function __construct(Router $router = null)
     {
@@ -75,18 +73,12 @@ class AuraRouter implements RouterInterface
         $this->router = $router;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function addRoute(Route $route)
+    public function addRoute(Route $route) : void
     {
         $this->routesToInject[] = $route;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function match(Request $request)
+    public function match(Request $request) : RouteResult
     {
         // Must inject routes prior to matching.
         $this->injectRoutes();
@@ -98,22 +90,10 @@ class AuraRouter implements RouterInterface
             return $this->marshalFailedRoute($request, $matcher->getFailedRoute());
         }
 
-        // The $allows property is empty if no HTTP methods were specified
-        // during route creation; such a situation is a (potential) 405.
-        // We have to retrieve the value first, because Aura\Route uses
-        // property overloading, which does not play well with empty().
-        $allows = $route->allows;
-        if (empty($allows)) {
-            return $this->handleRouteWithUndefinedHttpMethods($route, $request);
-        }
-
         return $this->marshalMatchedRoute($route);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function generateUri($name, array $substitutions = [], array $options = [])
+    public function generateUri(string $name, array $substitutions = [], array $options = []) : string
     {
         // Must inject routes prior to generating URIs.
         $this->injectRoutes();
@@ -123,10 +103,8 @@ class AuraRouter implements RouterInterface
 
     /**
      * Create a default Aura router instance
-     *
-     * @return Router
      */
-    private function createRouter()
+    private function createRouter() : Router
     {
         return new Router();
     }
@@ -136,30 +114,23 @@ class AuraRouter implements RouterInterface
      *
      * If the route failure is due to the HTTP method, passes the allowed
      * methods when creating the result.
-     *
-     * @param Request $request
-     * @param null|AuraRoute $failedRoute
-     * @return RouteResult
      */
-    private function marshalFailedRoute(Request $request, AuraRoute $failedRoute = null)
-    {
+    private function marshalFailedRoute(
+        Request $request,
+        AuraRoute $failedRoute = null
+    ) : RouteResult {
         // Evidently, getFailedRoute() can sometimes return null; these are 404
         // conditions. Additionally, if the failure is due to inability to
         // match the path, that to is a 404 condition.
         if (null === $failedRoute
             || $failedRoute->failedRule === PathRule::class
         ) {
-            return RouteResult::fromRouteFailure();
-        }
-
-        // Allow HEAD and OPTIONS requests if the failed route matches the path
-        if (in_array($request->getMethod(), self::HTTP_METHODS_IMPLICIT, true)) {
-            return $this->marshalMatchedRoute($failedRoute);
+            return RouteResult::fromRouteFailure(Route::HTTP_METHOD_ANY);
         }
 
         // Check to see if we have an entry in the method path map; if so,
         // register a 405 using that value.
-        list($path) = explode('^', $failedRoute->name);
+        [$path] = explode('^', (string) $failedRoute->name);
         if (array_key_exists($path, $this->pathMethodMap)) {
             return RouteResult::fromRouteFailure($this->pathMethodMap[$path]);
         }
@@ -173,21 +144,18 @@ class AuraRouter implements RouterInterface
             return RouteResult::fromRouteFailure($failedRoute->allows);
         }
 
-        return RouteResult::fromRouteFailure();
+        return RouteResult::fromRouteFailure($failedRoute->allows ?: Route::HTTP_METHOD_ANY);
     }
 
     /**
      * Marshals a route result based on the matched AuraRoute and request method.
-     *
-     * @param AuraRoute $auraRoute
-     * @return RouteResult
      */
-    private function marshalMatchedRoute(AuraRoute $auraRoute)
+    private function marshalMatchedRoute(AuraRoute $auraRoute) : RouteResult
     {
         $route = $this->matchAuraRouteToRoute($auraRoute);
         if (! $route) {
             // This should likely never occur, but is present for completeness.
-            return RouteResult::fromRouteFailure();
+            return RouteResult::fromRouteFailure(Route::HTTP_METHOD_ANY);
         }
 
         return RouteResult::fromRoute($route, $auraRoute->attributes);
@@ -196,7 +164,7 @@ class AuraRouter implements RouterInterface
     /**
      * Loops through any un-injected routes and injects them into the Aura.Router instance.
      */
-    private function injectRoutes()
+    private function injectRoutes() : void
     {
         foreach ($this->routesToInject as $index => $route) {
             $this->injectRoute($route);
@@ -207,10 +175,8 @@ class AuraRouter implements RouterInterface
 
     /**
      * Inject a route into the underlying Aura.Router instance.
-     *
-     * @param Route $route
      */
-    private function injectRoute(Route $route)
+    private function injectRoute(Route $route) : void
     {
         $path = $route->getPath();
 
@@ -270,40 +236,5 @@ class AuraRouter implements RouterInterface
 
             return false;
         }, false);
-    }
-
-    /**
-     * Handle a route with undefined allowed HTTP methods.
-     *
-     * Aura\Route::$allows can be empty in one of two situations:
-     *
-     * - all HTTP methods are supported
-     * - no HTTP methods are supported
-     *
-     * These need to be handled differently, so this method attempts to retrieve
-     * the associated Zend\Expressive\Router\Route instance.
-     *
-     * If found, it checks to see if the route allows ANY HTTP method, and, if
-     * so, marshals a successful route result.
-     *
-     * Otherwise, it marshals a failed route result (contingent on implicit
-     * support for HEAD and OPTIONS).
-     *
-     * @param AuraRoute $auraRoute
-     * @param Request $request
-     * @return RouteResult
-     */
-    private function handleRouteWithUndefinedHttpMethods(AuraRoute $auraRoute, Request $request)
-    {
-        $route = $this->matchAuraRouteToRoute($auraRoute);
-        if (! $route) {
-            return $this->marshalFailedRoute($request, $auraRoute);
-        }
-
-        if ($route->getAllowedMethods() === Route::HTTP_METHOD_ANY) {
-            return $this->marshalMatchedRoute($auraRoute);
-        }
-
-        return $this->marshalFailedRoute($request, $auraRoute);
     }
 }